refactor: clean up operator transformers
